+++ stories/BrandAssets.stories.ts
@@ -1,0 +1,11 @@
+import BrandAssets from './BrandAssets.svelte';
+
+export default {
+  title: 'Design System/Brand Assets',
+  component: BrandAssets,
+  parameters: {
+    layout: 'fullscreen'
+  }
+};
+
+export const FallbackSet = {};